Move Game management logic to own module
