+++ haverer-api.cabal
@@ -23,6 +23,8 @@ library
                      Handler.Common
                      Handler.Game
                      Handler.Home
+                     GameManagement
+
 
     if flag(dev) || flag(library-only)
         cpp-options:   -DDEVELOPMENT
